[UPDATE] pipeline 47

+++ Jenkinsfile
@@ -39,12 +39,20 @@ pipeline {
                     //         exit 1
                     //     fi
                     // ''', returnStatus: true) == 0
+                    def imageId = sh(script: 'docker images -q soa-deploy:latest', returnStdout: true).trim()
+                    if (imageId != "") {
+                        def contenerdorId = sh(script: 'docker ps -q -f name=soa-deploy-test', returnStdout: true).trim()
+                        if (contenerdorId != "") {
+                            sh "docker stop soa-deploy-test"
+                            sh "docker rm soa-deploy-test"
+                            // sh "docker rmi ${imageId}"
+                            sh "docker rmi soa-deploy:latest"
+                        } else {
+                            sh "docker rmi soa-deploy:latest"
+                        }
+                    }
                     echo 'Building image docker...'
                     sh "docker build -t soa-deploy:latest ."
-                    def imageId = sh(script: 'docker images -q :', returnStdout: true).trim()
-                    if (imageId != "") {
-                        echo "${imageId}"
-                    }
                 }
             }
         }
@@ -88,25 +96,18 @@ pipeline {
                     // sh "docker ps -q -f name=soa-deploy-test"
                     def contenerdorId = sh(script: 'docker ps -q -f name=soa-deploy-test', returnStdout: true).trim()
                     echo "Contenedor eliminado: ${contenerdorId}"
-                    // if (containerRunning) {
-                    //     echo 'Actualizando contenedor...'
-                    //     // Si el contenedor ya está corriendo, actualiza la imagen
-                    //     sh "sudo docker stop soa-deploy-test"
-                    //     sh "sudo docker rm soa-deploy-test"
-                    // }
+                    if (containerRunning != "") {
+                        sh "docker stop soa-deploy-test"
+                        sh "docker rm soa-deploy-test"
+                    }
                     // echo 'Desplegando...'
                     // Inicia el contenedor con la nueva imagen
                     // 1c17aed9e12545ecb784479826baae18bd30424a36a946f3133a11ed798ec537 = soa-deploy-test
                     // sh "docker stop 1c17aed9e12545ecb784479826baae18bd30424a36a946f3133a11ed798ec537"
                     // sh "docker rm 1c17aed9e12545ecb784479826baae18bd30424a36a946f3133a11ed798ec537"
                     // sh "docker run -d -p 3000:3000 --name 1c17aed9e12545ecb784479826baae18bd30424a36a946f3133a11ed798ec537 soa-deploy:latest"
-                    sh "docker stop soa-deploy-test"
-                    sh "docker rm soa-deploy-test"
-                    def imageId = sh(script: 'docker images -q <none>:<none>', returnStdout: true).trim()
-                    if (imageId != "") {
-                        echo "id imagen anterior ${imageId}"
-                    }
-                    sh "docker rmi ${imageId}"
+                    // sh "docker stop soa-deploy-test"
+                    // sh "docker rm soa-deploy-test"
                     sh "docker run -d -p 3000:3000 --name soa-deploy-test soa-deploy:latest"
                 }
             }
